+++ mrv2/lib/mrvWidgets/mrvVersion.h
@@ -23,11 +23,6 @@ namespace mrv
 
     std::string about_message();
 
-    void memory_information(
-        uint64_t& totalVirtualMem, uint64_t& virtualMemUsed,
-        uint64_t& virtualMemUsedByMe, uint64_t& totalPhysMem,
-        uint64_t& physMemUsed, uint64_t& physMemUsedByMe);
-
     std::string cpu_information();
     std::string gpu_information(ViewerUI* uiMain);